refactor(mcp): split registered lifecycle slice out of extension_lifecycle
Behavior-preserving move: owner-scope helpers, live registered
resolution, tenant-match guard, and restore's per-owner-batched
registered fallback go to extension_lifecycle/registered_lifecycle.rs
(install_policy.rs precedent); the owner-registered tests move to
extension_lifecycle/tests/registered_lifecycle_tests.rs. Visibility
widened to pub(super) only where the move required it.
